Moved inputed into rsky-lexicon package
