+++ rsky-lexicon/src/com/atproto/identity.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use std::collections::BTreeMap;
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct ResolveHandleOutput {
@@ -11,4 +12,14 @@ pub struct ResolveHandleOutput {
 pub struct UpdateHandleInput {
     /// The new handle.
     pub handle: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SignPlcOperationRequest {
+    pub token: String,
+    pub rotation_keys: Vec<String>,
+    pub also_known_as: Vec<String>,
+    pub verification_methods: BTreeMap<String, String>,
+    pub services: BTreeMap<String, Service>,
 }
